app-level insert notifications, rework scheduler
Instead of emitting notifications for all inserts with a database
trigger, use application-level notification logic with a per-queue
debouncer. This enables us to drastically reduce the amount of
notifications emitted to 1 per inserting client per cooldown interval
per queue.

The `JobScheduler` has been reworked to utilize this same mechanism. In
addition, the scheduler has been updated to use a "look ahead" concept,
meaning any jobs that are scheduled before its next planned run should
be preemptively marked as available (while preserving their
`ScheduledAt` time). This makes jobs available for scheduled work sooner
and without delay, with the caveat that a notification may not be
emitted right when a job is scheduled. On queues without any ongoing
insert activity, this could potentially increase the latency before a
scheduled job gets picked up by a worker, which is a worthwhile tradeoff
for the performance optimization we gain on high throughput queues. This
can be counteracted if desired by lowering the client's polling
interval.