Take a stab at fixing test with intermittent `i/o timeout` problem
A first pass at trying to fix the problem described in #289, in which
what seems to be a finishing test produces an `i/o timeout` while trying
to insert a job.

I can't tell for sure, and this is another problem ~impossible to repro
locally, but I'm guessing the problem is that a cancelled context may
bubble up as a number of underlying problems in pgx, and this failure is
occurring because we're only checking against `context.Canceled`.

Here, modify the approach to check on `ctx.Err()` instead, and ignore
any insert problems if the parent context is indeed cancelled.

Fixes #289.