remove inefficient created, attempted filters. Limits on finalized
River does not keep indexes on `created_at` and `attempted_at` fields
and we do not want to do so for performance reasons. We shouldn't offer
a built-in API for sorting by these fields which will be guaranteed to
have poor performance on large tables.

Additionally, while `finalized_at` _is_ indexed, it's indexed via a
_partial_ index on non-null values (i.e. only for finalized jobs). We
can still make it possible to order based upon this field, but only when
also filtering to only finalized states. This still _may_ be inefficient
at times, but it should hopefully be able to at least avoid a full table
scan.