Prepare release `v0.3.0` (#303)
Prep release of `v0.3.0`. This is a relatively small one containing
mainly #281 and #297, but I think it's not a bad idea to cut a release
since we haven't done one in a while, and to get these additive changes
out ahead of the number of breaking changes that'll probably come with
the next release.