Replace list of all job states with `rivertype.JobStates()`
A small change that removes a list of all job states defined in
`insert_opts.go` that was used to validate states sent into `UniqueOpts`
with `rivertype.JobStates()` which was introduced recently n #297.
Removing one of the lists means we only need to maintain one of them.

I left the variable in place instead of invoking `rivertype.JobStates()`
so that we don't need to initialize a new slice every time we validate.