Fix bug in `SharedTx` when `Query` returns on error condition (#311)
Fix a hard-to-spot but long running bug in `SharedTx` in which when
`Query` was invoked, the standard success case unlocked the `SharedTx`
as expected, but if returning under error condition, the parent
`SharedTx` would not be unlocked.

This wasn't firing most of the time, but could occur in some stress
tests where maintenance services would occasionally successfully make it
into their work loop before their context was cancelled. The context
cancellation would error `Query`, not unlock `SharedTx`, and cause an
error if another service then later on entered its work loop and tried
to use `SharedTx`.