Prepare version `v0.4.1` (#318)
A small release containing the change in #315. I don't _think_ anything
critical was broken, but not completely certain, and a release is an
easy thing to do, so just in case.

Also update `./cmd/river`'s references to `v0.4.0` which shouldn't
affect anything, but a minor nicety to stay current.