Add subscription events that are emitted when a queue is paused or resumed
Here, largely as a potential way to resolve #324, add new events
`EventKindQueuePaused` and `EventKindQueueResumed` which are emitted
whenever a queue is paused or resumed. Tests will be able to use these
to wait on the completion of these actions even if calls to `QueuePause`
or `QueueResume` are asynchronous (as they are currently), or if the
pause/resume was initiated by a different client. I expect them to also
be useful for general telemetry/logging purposes.

Recall that when we first designed the subscription API, we explicitly
_did not_ add a way to subscribe to all possible event types. This means
that although the new queue events fill `Event.Queue` instead of
`Event.Job` and `Event.JobStats`, the API is still forward compatible
because nobody could be listening to these yet.