Wrap queues returned from list in result type like we do for rows (#326)
Something I noticed while rereading some of the queue code is that
although we added a `QueueListParams` to make queue listing symmetric
with job listing (and leaving space for future filters and a possible
cursor), we were still returning `[]*Queue`, which has the downside of
not leaving any space for a last cursor property should we add one in
the future.

Here, in the interest of keeping job/queue listing looking like each
other, have `QueueList` return a `QueueListResult`. So far this struct
contains only a `Queues` property, but a cursor could be added to it
should this functionality be implemented later.