Improve documentation for insert-only clients (#375)
Try to address #371 in that it's currently too hard to discover that an
insert-only client is possible, and how to initialize one. There is some
documentation on the `Queues` variable, but it's not that easy to find.

Here, add docs in a few other places under the principle that some docs
repetition isn't the worse thing in that it helps make these concepts
more obvious to find. We put docs in the README and package docs (which
mirror each other), along with some on the `river.Config` struct.

Fixes #371.