normalize subscription manager service
Respect a stop, but in the case of one, makes sure to clear the
subscription channel before leaving, which means that it still correctly
clears all events on a client shutdown. This gives us a way to use the
stress test because all we need to do is close the channel in advance
before calling startstoptest.Stress (the service is still a little weird
compared to other because it requires that channel close, but a little
less so).

Also normalizes things a bit by removing the custom Stop implementation,
which most services shouldn't need.

Co-Authored-By: Brandur Leach <brandur@brandur.org>