Fix broken links from Godoc
I just happened to be looking at our Godoc and noticed that some of the
links are broken [1]. Some were missing their definitions, but the new
River UI ones use a Markdown syntax that's not supported by Godoc.

I also updated the README with the same changes. They're not necessary
there, but IMO better to keep these two docs as synchronized as possible
so if we need to make wider changes later it's easy to change both.

[1] https://pkg.go.dev/github.com/riverqueue/river